refactor(scripts): ship fpga_init/observe as wheel-installed console scripts
Move the SNAP supervisor and HDF5 writer from scripts/ into the
package at src/eigsep_observing/scripts/, expose them via
[project.scripts] as eigsep-fpga-init and eigsep-observe, and point
the systemd units at the venv-resolved binaries. Required by
downstream image builds (eigsep-field) that install from a
wheelhouse and have no source checkout, so scripts/ on disk is not
available to ExecStart=.

Also lazy-imports IPython inside --interactive so the production
supervisor service does not pull in the [interactive] extra. Updates
operator-facing docstrings, error messages, and OPERATIONS.md to use
the console-script names.
