+++ tests/test_live_status_aggregator.py
@@ -453,9 +453,9 @@ def test_snap_tick_reads_file_heartbeat_from_redis(agg, seeded):
 
 
 def test_snap_tick_reads_snap_reinit_from_redis(agg, seeded):
-    """fpga_init.py publishes a reinit heartbeat on each supervised
-    --reinit; the aggregator must surface it on the same SNAP-side
-    transport so the dashboard can render the count."""
+    """eigsep-fpga-init publishes a reinit heartbeat on each
+    supervised --reinit; the aggregator must surface it on the same
+    SNAP-side transport so the dashboard can render the count."""
     from eigsep_observing.snap_reinit import publish as publish_reinit
 
     snap, _ = seeded
@@ -471,8 +471,8 @@ def test_snap_tick_reads_snap_reinit_from_redis(agg, seeded):
 
 
 def test_snap_tick_snap_reinit_absent_returns_empty(agg, seeded):
-    """When fpga_init.py has never published (e.g. fresh deploy or
-    Redis was flushed), the aggregator carries the empty-sentinel
+    """When eigsep-fpga-init has never published (e.g. fresh deploy
+    or Redis was flushed), the aggregator carries the empty-sentinel
     dict so the dashboard renders an unknown tile, not an error."""
     agg._snap_tick()
